requirements.txt
Django>=6.0
django-allauth[socialaccount]
django-crispy-forms
djangorestframework
Pillow
python-dotenv
